Set status message for ratelimit error responses

+++ synapse/handlers/_base.py
@@ -39,7 +39,7 @@ class BaseHandler(object):
         )
         if not allowed:
             raise LimitExceededError(
-                retry_after_ms=1000*(time_allowed - time_now),
+                retry_after_ms=int(1000*(time_allowed - time_now)),
             )
